requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-appconf==1.0.3
django-bootstrap-form==3.4
django-sendgrid-v5==0.8.0
djangorestframework==3.9.4
future==0.17.1
httpie==1.0.2
idna==2.8
pinax-notifications==5.0.3
pinax-templates==2.0.2
Pygments==2.4.0
python-http-client==3.1.0
pytz==2019.1
requests==2.22.0
sendgrid==6.0.5
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.2
